refactor(go): clean graph memory owner sizing

+++ go/fory/field_serializer.go
@@ -68,7 +68,7 @@ func newDeclaredSliceSerializer(type_ reflect.Type, elemSerializer Serializer, r
 	if elem.Kind() == reflect.Ptr && elem.Elem().Kind() == reflect.Interface {
 		return nil, fmt.Errorf("slice serializer does not support pointer to interface element type: %v", type_)
 	}
-	elemBytes := int64(elem.Size())
+	elemBytes := int(elem.Size())
 	return &sliceSerializer{
 		type_:          type_,
 		elemSerializer: elemSerializer,
